amazon-spa/cdk/requirements.txt
-e .
aws-cdk.core==1.143.0
aws-cdk.aws-iam==1.143.0
aws-cdk.aws-s3==1.143.0
aws-cdk.aws-cloudfront==1.143.0
aws-cdk.aws-cloudfront-origins==1.143.0
aws-cdk.aws-logs==1.143.0
aws-cdk.aws-certificatemanager==1.143.0
aws-cdk.aws-lambda==1.143.0